XChangeProperty() expects values of size 32 to be longs (see "man XChangeProperty"). Changed parameter types to long when calling XChangeProperty due to https://bugs.freedesktop.org/show_bug.cgi?id=16802
BUG=296940
R=tony@chromium.org,erg@chromium.org

Review URL: https://codereview.chromium.org/24281007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@226245 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d7a7c8066134dc6e736ee8a16117cd66c6b8d92e

+++ ui/base/x/x11_util.cc
@@ -553,7 +553,8 @@ void* GetVisualFromGtkWidget(GtkWidget* widget) {
 
 void SetHideTitlebarWhenMaximizedProperty(XID window,
                                           HideTitlebarWhenMaximized property) {
-  uint32 hide = property;
+  // XChangeProperty() expects "hide" to be long.
+  unsigned long hide = property;
   XChangeProperty(gfx::GetXDisplay(),
       window,
       GetAtom("_GTK_HIDE_TITLEBAR_WHEN_MAXIMIZED"),
